+++ src/main/java/blackjack/BlackJackApplication.java
@@ -4,7 +4,7 @@ import blackjack.domain.game.BlackJackGame;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Deck;
 import blackjack.domain.gamer.Gamers;
-import blackjack.dto.GameResultDto;
+import blackjack.domain.game.BlackJackReferee;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
 
@@ -19,7 +19,7 @@ public class BlackJackApplication {
         Gamers gamers = blackJackGame.getGamers();
         OutputView.printGamers(gamers.getDealer(), gamers.getPlayers());
 
-        GameResultDto result = blackJackGame.play(InputView::askHitOrStay, OutputView::checkHoldingCards);
+        BlackJackReferee result = blackJackGame.play(InputView::askHitOrStay, OutputView::checkHoldingCards);
 
         OutputView.printAdditionalDrawDealer(result.getDealerDrawCount());
         OutputView.printFinalCards(gamers.getDealer(), gamers.getPlayers());
